# Conflicts:
#	R.NET/RDotNet.csproj
#	RDotNet.Graphics/RDotNet.Graphics.csproj
#	RDotNet.NativeLibrary/RDotNet.NativeLibrary.csproj

+++ R.NET/Internals/SEXPREC.cs
@@ -82,11 +82,13 @@ namespace RDotNet.Internals
         internal promsxp promsxp;
     }
 
+    // In R 3.5, the length & true length values went from 32 to 64 bits in length.  These are defined in R as R_xlen_t
+    // (previously R_len_t) - https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h
     [StructLayout(LayoutKind.Sequential)]
     internal struct vecsxp
     {
-        public int length;
-        public int truelength;
+        public long length;
+        public long truelength;
     }
 
     [StructLayout(LayoutKind.Sequential)]
